requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
torch>=2.0
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
umap-learn>=0.5
scipy>=1.11
jupyter>=1.0
ipykernel>=6.0
